Refactor history & staking methods
